requirements.txt
asgiref==3.8.1
cfgv==3.4.0
coverage==7.6.4
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
filelock==3.16.1
identify==2.6.1
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.7.1
sqlparse==0.5.1
virtualenv==20.27.0
